Correct return behavior of mutex_lock_interruptible()
The model was returning a nondeterministic value regardless of whether
locking was successful.  According to the spec,
(https://www.kernel.org/doc/htmldocs/kernel-locking/API-mutex-lock-interruptible.html),
the return value should be zero if the lock aquisition was successful,
and one otherwise.

The model now chooses a return value, and completes the lock if the
return value was 0.